requirements.txt
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.7.0
pypdf==4.2.0
python-dotenv==1.0.1
google-genai==0.3.0
tqdm==4.66.4
torch==2.2.2
scikit-learn==1.4.2
scipy==1.13.0
tqdm==4.66.4
streamlit==1.36.0
